Use interior mutability in generation_list_t
The previous approach transmuting the values to `Cell<u64>` was extremely
unsound since the compiler doesn't treat structs containg `Cell<T>` the same way
that it treats `T`.

It wasn't possible with the ffi because `at_mut()` returns a mutable reference
and we really don't want to use `RefCell` here when the usecase is practically
crying for `Cell` instead.